Added Hash ID library : can encode ids with no need for db lookup
Added the library (https://github.com/ivanakimov/hashids.php)
through composer, and then made an example and
test file to show how different scripts in different places can encode
and decode ids without using the database.

Added a new file to include a salt phrase and padding that can be used
accross different projects. The padding and salt can be changed,
but all the scripts on the server should use the same ones

The test and example script is at
scratch/test_hash_ids.php

The length of the encoded string can be adjusted, but
 here everything is set to 8 characters

Output of the test file (will always be the same with these settings)

1 is encoded to NG0aX8B5
NG0aX8B5 is decoded to 1
2 is encoded to vNoNMQLr
vNoNMQLr is decoded to 2
100321 is encoded to lo3vy6vQ
lo3vy6vQ is decoded to 100321
9009321 is encoded to o3wBjldo
o3wBjldo is decoded to 9009321
89009321 is encoded to o3Jz559Q
o3Jz559Q is decoded to 89009321

Also, these are designed to be put into urls.
So these will be url friendly, including no vulgar words being accidentally generated.
These are very much youtube ids