Created database table migrations for test suits. Updated phinx
Updated Phix to 8.1
phinx is now part of cakephp, but I use it as a stand alone in the transcription project.
The new version adds more features
and the change log is at https://github.com/cakephp/phinx/blob/master/CHANGELOG.md
So, I replaced the older phar with the new one in this directory.

The steps to do this is as follows (in case you want to do add it elsewhere, you can just copy the bin/phinx.phar in the project ).
Also, these steps can be repeated with future version upgrades. Just change the branch arg in the clone command below

* go outside the directories of this project and make a folder you can delete after all this
* go inside that new folder
* git clone https://github.com/cakephp/phinx.git --branch v0.9.1
* cd phinx
* php composer.phar install  (I have composer set up to run on the command line)
* curl -LSs https://box-project.github.io/box2/installer.php | php (the software to make this a phar)
* php box.phar build

this will make a phinx-v0.9.1.phar, rename it to phinx.phar, give it execute privledge, and move  where needed, delete temp directory
the version when run will say 0.8.1 (older one was 0.6.4)

Created Database Tables for the tests (6 tables)

Workflow is as follows

1) when a test is run, it finds itself in the database, in the table tests. If cannot find itself exits
2) It fills out and saves a new test_runs, and includes the starting time
3) If there are any resources (user or transaction), it finds them in the seeds table
4) It marks any resources as locked in the test_resource_locks. By making and saving a new row.If cannot get a lock then exits
5) It pulls in the seed data for each resource and updates all the fields each resource
6) The test is then run
7) The result of the test (console message and pass fail status), is saved to the test_runs it created, as well as the end time
8) Any resources it locked is unlocked by putting in a lock_time_end value

Later on, we can have a cron job scanning the test_runs table, and any production test that fails can trigger a notice

The tables are in phinx migration format, all the columns are commented.
And there are some triggers to make sure the enumerations are kept accurate, to prevent typos
All tables are in mb4 mysql and maria unicode format
I added configs for the server to allow it to be mb4 also, if needed